fix:Updated palindromePartition algorithm
